refact : 탈퇴기능 이동
auth -> user 로 도메인 변경
    탈퇴는 인증/인가가 아닌 유저에 해당하는 기능으로 판단
