+++ src/main/java/com/Soo_Shinsa/service/AuthService.java
@@ -12,5 +12,4 @@ public interface AuthService {
 
     JwtAuthResponseDto login(LoginRequestDto dto);
 
-    void leave(String password, User user);
 }
