+++ src/main/java/com/Soo_Shinsa/service/UserService.java
@@ -3,9 +3,12 @@ package com.Soo_Shinsa.service;
 import com.Soo_Shinsa.dto.UserDetailResponseDto;
 import com.Soo_Shinsa.dto.UserUpdateRequestDto;
 import com.Soo_Shinsa.model.User;
+import jakarta.validation.constraints.NotBlank;
 
 public interface UserService {
     UserDetailResponseDto getUser(User user);
 
     UserDetailResponseDto updateUser(User user, UserUpdateRequestDto userUpdateRequestDto);
+
+    void leave(@NotBlank String password, User user);
 }
